Update zfs share to support read and write ip
This commit update zfs share to support setting a read ip default value,
read, and write IP addresses separately. Without this change it is not
possible to grant access to a default read ip, along with separate read
and write IP addresses. Additionally, nosub and anon options are added
to the share define.