Remove guestok parameter
This commit removes the guestok parameter, which is a duplicate of the share_smb_guestok parameter. Without this change there is a parameter in the type that is a duplicate and not implemented in the provider.

Additionally, share_auto default is updated from ('on') to :on.
Share_nfs default is updated to be :on for the sake of consistency.