Check sync before snapshot
This commit adds a check in zfs_snapshot to determine whether a sync is in progress on the target volume. If any file named *.lock exists on the root level of the target volume then the task will wait as long as 12 hours, checking every 10 seconds, to determine if a sync has completed before creating a snapshot. Without this change it's possible for a snapshot to be created during a file sync, which makes its results less predictable.